requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.2
django-extensions==4.1
djangorestframework==3.16.0
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0

openai~=1.86.0
